chore(cargo)!: Replace envy and shellexpand with directories
The directories dependency does what I was going to do with envy and shellexpand
for finding and setting up Ricer's configuration files by hand. The directories
dependency also abides by the XDG base directory standard[1] on its own, which
was what I was going to implement on my own. Now I do not need to reinvent the
wheel :^)

[1] https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html

Signed-off-by: Jason Pena <jasonpena@awkless.com>