feat(ricer::cli)!: Use external command for using Git binary on a repository
Originally I was going to just use git2-rs to implement a portion of the Git
command set to manipulate user repositories. However, I realized that I want to
have the functionality to run the full Git command set on a repository through
Ricer rather than just using the enter command to enter a target repository to
execute Git in a sub-shell.

I essentially want a basic shortcut like this:

	ricer <REPO> <GIT_CMD>

I will keep the enter command just in case I want to have the ability to enter
a target repository directly.

Signed-off-by: Jason Pena <jasonpena@awkless.com>