ref(ricer::cli)!: Separate hook execution into individual options
Originally, I was going to have `run_hooks` control how both command hooks and
repository hooks are executed. However, I realized that this was a bad idea,
because I was making it seem that both hook types were the same.

Thus, I split the execution options for both hook types into their own little
options to make sure that the user understands that they are distinct, and to
give them more flexibility in how these hooks should be executed!

Signed-off-by: Jason Pena <jasonpena@awkless.com>