ref(ricer::config)!: Improve config entry implementation
BREAKING CHANGES: Moved out file and directory manager code out. Will be placing
this manager code in a separate module. The `*Entry` API and their associated
builder patterns have been simplified with improved documentation. The API for
each `*Entry` object has been improved to also include additional fields to
improve existing configuration functionality in the future for the user and
Ricer's command set as a whole.

More specifically, the `RepoEntry` API has been extended to include
bootstrapping options for the user. The `CmdHookEntry` API removes the `repo`
field for `workdir` field to make it easier for the user to control the
working directories of their hook scripts without the need to rely on specific
repository definitions. This also decouples `CmdHookEntry` from `RepoEntry`,
which will make the implementation of the repository manager and hook manager
simpler and avoid mutability issues compared to the original design.

Signed-off-by: Jason Pena <jasonpena@awkless.com>