feat(ricer::ui::CmdSet): add git as external subcommand
Defined shortcut command for Ricer. This shortcut will take the form of:

	# ricer <REPO> <GIT_CMD>

Where <REPO> is the name of a target repository, and <GIT_CMD> is a Git command
to run on <REPO> itself. This shortcut eliminates the need to use the _enter_
subcommand that Ricer provides to allow the user to directly use Git on a given
repository.

Signed-off-by: Jason Pena <jasonpena@awkless.com>