feat(ricer::context): add `Context`
A _context_ is a definitive and flattened set of options required for a Ricer
command to function.

Various `[...]Opts` structs contained in `Cli` provide an interface for the
user. These structs provide a tree-like structure that can be annoying to use
internally. Thus, the context layer converts various options offered by the CLI
into a flattened struct with at most one to two levels of indirection.

Another benefit of the context layer, is the decoupling of Ricer commands from
the CLI. Thus, we can make modifications to the CLI freely without worring about
affecting any of the command set implementations.

Signed-off-by: Jason Pena <jasonpena@awkless.com>