refactor(ricer::config)!: combine `data_xchg` module and `ConfigManager` into `config` module
Upon some reflection of the current state of the codebase, I decided to combine
the `ricer::data_xchg` module and the logic for configuration file management in
the `ricer::manager` module together into one module named `ricer::config`.

It is clear that this code belongs together in one module, because their logic
shares the common goal of parsing and manipulating configuration file data for
the Ricer binary. Thus, it makes more logic sense to me to have them together
in _one_ module.

[BREAKING CHANGES]: place all configuration file logic into `ricer::config` module.
Signed-off-by: Jason Pena <jasonpena@awkless.com>