refactor(ricer::hook)!: remove `CmdHook::set_pager`
The `CmdHook::set_pager` method is not needed because I no longer intend to
support dependency injection for a pager. I can just disable the pager through
`ricer::context::HookAction` and manually test the pager on my own, because I do
not know any reliable way to automate the testing of `HookPager`.

Signed-off-by: Jason Pena <jasonpena@awkless.com>