feat(ricer::vcs): add test case for `GitRepo::init_fake_bare`
Red phase: get `GitRepo::init_fake_bare` to produce Self that uses fake bare
repository technique similar to how VCSH does it.

A fake bare repository pretends to be bare but isn't. The working directory is
separate from the git directory such that the git directory does not contain a
`.git/` folder similar to a bare repository. This allows us to have a git
repository with a separate working directory like a bare repository, but the
repository still keeps track of the working directory unlike a bare repository.

Signed-off-by: Jason Pena <jasonpena@awkless.com>